requirements_full.txt
alembic==1.11.3
anyio==4.0.0
appnope==0.1.3
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==23.7.0
bleach==6.0.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.7.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.3
executing==1.2.0
eyeD3==0.8.10
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Alembic==2.0.1
Flask-Migrate==4.0.4
Flask-Script==2.0.6
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fqdn==1.5.1
greenlet==2.0.2
grpcio==1.30.2
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
iniconfig==2.0.0
install==1.3.5
ipykernel==6.25.2
ipython==8.14.0
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
more-itertools==8.10.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nemo-emblems==5.8.0
nest-asyncio==1.5.7
netaddr==0.8.0
netifaces==0.11.0
notebook==7.0.3
notebook_shim==0.2.3
oauthlib==3.2.0
onboard==1.4.1
overrides==7.4.0
packaging==23.1
PAM==0.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.10.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.16.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.4.2
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-json-logger==2.0.7
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.2
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.20
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
uri-template==1.3.0
urllib3==1.26.5
virtualenv==20.24.4
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.3.7
WTForms==3.0.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
